Rename C++ extensions to .cpp and .hpp
This commit adds updates 2 features:
1. Renames .hh extensions to .hpp and .cc extensions to .cpp
2. Renames cmk.hh to charmlite.hpp

Random Flyby:
Add .gitignore to the project
