+++ examples/cdt/pgm.cpp
@@ -7,8 +7,8 @@
  * author: j. szaday <szaday2@illinois.edu>
  */
 
-#include <cmk.hh>
-#include <completion.hh>
+#include <charmlite.hpp>
+#include <completion.hpp>
 
 // a callback to resume the main thread
 void resume_main_(cmk::message_ptr<>&& msg);
