+++ include/charmlite.hpp
@@ -1,13 +1,13 @@
 #ifndef __CMK_HH__
 #define __CMK_HH__
 
-#include "collection.hh"
-#include "core.hh"
-#include "proxy.hh"
-#include "reduction.hh"
+#include "collection.hpp"
+#include "core.hpp"
+#include "proxy.hpp"
+#include "reduction.hpp"
 
 // ( no reordering )
-#include "cmk.impl.hh"
+#include "charmlite.impl.hpp"
 
 namespace cmk {
     // broadcasts an exit message to all pes
