+++ include/charmlite.impl.hpp
@@ -1,8 +1,8 @@
 #ifndef __CMK_IMPL_HH__
 #define __CMK_IMPL_HH__
 
-#include "ep.hh"
-#include "message.impl.hh"
+#include "ep.hpp"
+#include "message.impl.hpp"
 
 /* registers all user data-types with the RTS
  */
