+++ include/message.impl.hpp
@@ -1,7 +1,7 @@
 #ifndef __CMK_MESSAGE_IMPL_HH__
 #define __CMK_MESSAGE_IMPL_HH__
 
-#include "message.hh"
+#include "message.hpp"
 
 namespace cmk {
     template <typename T>
