+++ include/reduction.hpp
@@ -1,8 +1,8 @@
 #ifndef __CMK_REDUCTION_HH__
 #define __CMK_REDUCTION_HH__
 
-#include "callback.hh"
-#include "message.hh"
+#include "callback.hpp"
+#include "message.hpp"
 
 // TODO ( converse collectives should be isolated/removed )
